Change from particle list to particle vector
To better facilitate copying to GPU, use a vector instead of list.
This makes the sequential version much more inefficient (due to the
removal procedure, I believe), but will allow for direct memory copying
to the GPU.
Note, this will cause uncoalesced memory accesses on the GPU when
accessing the position and velocity structures, but this is planned to
be dealt with later.