Count # inv parts; chg dbg print to show switch
Using thrust to count the number of invalid particles, which can be used
to calculated the number of valid particles.
Use these values to display the transition from valid to invalid in the
sorted vector.
Proves actual sorting took place.
Interesting: the number of valid particles seems to always be 10150, or
90 less than 10240, which is 32*32*10.
Is this reflective of the statistical thermodynamics?
I am not sure.