Translate initCollision, including stuct changes
Need to pass in num_cells, so added to struct.
Passing in global variables requires adding to the function arguments.
Have to use one of the curand arrays I had, but this will be used before
the initParticles part.
The order of usage does not matter for reproducibility, since every
thread has its own random state variable.