Add mapping struct(s) to pass to collide()
Need to sort particles, perform reduction, and compute prefix scan
of this reduced set to get a mapping from cell to particles.
Also, added in debugging print statements/functions.

Sorting the data makes the reduce/scan possible.
I am unsure if this will make accesses more coalesced, and given the
random nature of the collisions, this is probably not the case.
I still need to add the prefix-sum / scan operation and pass this in via