Translate randomDir() into GPU and use in init()
I avoided this before because I did not have the gpu version of the
vect3d, yet.
Now is failry simple.
Still have to decompose the vector into its components for storage, but
works.
There is probably a better way to deal with the vectors, such as using a
zip_iterator, but optimization comes later...