Sets the predicted max # of particles required
Instead of worrying about resizing the CPU and GPU vectors (which may be
easily facilitated by Thrust; I don't know for sure), this uses the
output from the CPU run to give an upper bounds on the nubmer of spots
needed to be reserved for the particles vector(s).
Essentially, I just looked at the behavior and noted the upper limit
after the run was finished.
Performance may improve for the beginning iterations if using a tighter
bounds on array sizes, but I am unsure how the resizing will affect
 this and/or if it is even worth worrying about for now.