Fix bug where particles were not reflecting
vel is just a temporary variable.
Have to write directly to the array to save the value.
I figured this would result in more particles in the volume, but it
seems the particles now have a hard cap at <97% of the num_cells*mppc
limit.
Is this because the reflected particles are making it all the way back
to the entrance and leaving that way?
That would imply the collision is incorrect, but given my need for this
fix, it is a distinct possibility....