Add all graphics-related functions/variables
Taken mostly from
https://github.com/NVIDIA/cuda-samples/blob/master/Samples/simpleGL/simpleGL.cu
Draws the particles and the plate, with the option to draw the edges
of the cells if the #if 0 around it is removed...
The main issue was the fact that for cell number values of 32x32x32,
the edges clutter the screen and do not look good. Only would be good
if could color by the concentration in the cell  and even then it would
almost be better to just color the particles in the cells instead.
Or do some kind of textured output with interpolations between cell
centers or something?
Ultimately, this looks okay for now, but needs significant design
improvement.
My not being an experienced graphics/simulation programmer grealty
limits my knowledge surrounding proper structures typically present in
these type of applications (i.e. I do not know what I am doing).
But it does "work". ¯\_(ツ)_/¯